# Perf fix vote collection