Address performance issue related to set_meta_votes.

The first 2 commits apply smaller updates close to refactoring, but they do change behavior:
-First use references to vector instead of cloning the contents.
-Then: like for parent_meta_votes ignore empty meta_votes everywhere meta_votes was called: rename to populated_meta_votes

Finally Update code to avoid duplicate lookup highlighted by 'perf' and 'callgrind'
```
RUSTFLAGS=-g RUST_BACKTRACE=1 cargo bench --features=testing -- --baseline=master-da5b291

minimal                 time:   [726.91 us 744.71 us 764.01 us]
                        change: [-17.884% -14.460% -10.985%] (p = 0.00 < 0.05)
                        Performance has improved.

static                  time:   [3.7173 ms 3.7292 ms 3.7455 ms]
                        change: [-12.445% -12.002% -11.570%] (p = 0.00 < 0.05)
                        Performance has improved.

dynamic                 time:   [2.4579 ms 2.4616 ms 2.4669 ms]
                        change: [-9.6652% -8.6950% -7.4378%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node4_opaque_evt8     time:   [1.4799 ms 1.4824 ms 1.4860 ms]
                        change: [-9.7467% -8.8776% -7.9721%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node8_opaque_evt8     time:   [8.3695 ms 8.4027 ms 8.4609 ms]
                        change: [-25.324% -24.511% -23.396%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node12_opaque_evt8    time:   [31.629 ms 31.786 ms 31.915 ms]
                        change: [-32.982% -32.524% -32.065%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node16_opaque_evt8    time:   [59.365 ms 59.632 ms 59.896 ms]
                        change: [-38.843% -38.333% -37.779%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node24_opaque_evt8    time:   [207.65 ms 209.20 ms 210.63 ms]
                        change: [-44.275% -43.856% -43.450%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node32_opaque_evt8    time:   [683.51 ms 707.38 ms 732.72 ms]
                        change: [-53.254% -52.011% -50.657%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node4_opaque_evt16    time:   [2.9375 ms 3.0356 ms 3.1498 ms]
                        change: [-13.487% -8.0745% -2.7411%] (p = 0.02 < 0.05)
                        Performance has improved.

a_node8_opaque_evt16    time:   [21.749 ms 21.877 ms 22.117 ms]
                        change: [-26.768% -24.192% -21.836%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node12_opaque_evt16   time:   [70.469 ms 70.733 ms 71.084 ms]
                        change: [-33.807% -30.776% -28.163%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node16_opaque_evt16   time:   [166.09 ms 166.51 ms 166.89 ms]
                        change: [-35.255% -34.316% -33.516%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node24_opaque_evt16   time:   [540.04 ms 544.97 ms 552.49 ms]
                        change: [-36.192% -35.299% -34.419%] (p = 0.00 < 0.05)
                        Performance has improved.

a_node32_opaque_evt16   time:   [1.3376 s 1.3637 s 1.3879 s]
                        change: [-39.308% -38.306% -37.251%] (p = 0.00 < 0.05)
                        Performance has improved.
```